Demo App: Webpack Build for Webpack 4
I wrote a new webpack configuration to be able to use the
demo app within the new toolchain.
I also removed some obscure resolves and fixed Webpack HMR.

The current version does not work with zepto.
https://github.com/madrobby/zepto/issues/1324
I am not willing to do this workaround. Maybe, there will
be a zepto version in near future.
