+++ package.json
@@ -29,9 +29,8 @@
     "clean": "npm-run-all clean:*",
     "clean:dist": "del-cli dist",
     "lint:autofix": "eslint src/js_es6 --fix",
-    "prestart": "npm-run-all --parallel clean",
-    "start": "npm run clean && cross-env JQUERY=true webpack --progress --env.dev --config webpack.config.dev.babel.js",
-    "start_es6z": "npm run clean && cross-env ZEPTO=true webpack --progress --env.dev --config webpack.config.dev.babel.js",
+    "start-jquery": "cross-env JQUERY=true webpack-dev-server --env.dev --config ./webpack/demo/webpack.config.js",
+    "start-zepto": "cross-env ZEPTO=true webpack-dev-server --env.dev --config ./webpack/demo/webpack.config.js",
     "test": "cross-env JQUERY=true karma start",
     "coverage": "JQUERY=true karma start --single-run --no-auto-watch --coverage=true || echo done"
   },
@@ -56,7 +55,7 @@
     "eslint-plugin-react": "^7.3.0",
     "file-loader": "^1.1.6",
     "html-loader": "^0.5.0",
-    "html-webpack-plugin": "^2.28.0",
+    "html-webpack-plugin": "^3.2.0",
     "image-webpack-loader": "^4.0.0",
     "jasmine": "^2.5.3",
     "jasmine-core": "^2.5.2",
@@ -82,7 +81,7 @@
     "url-loader": "^0.6.2",
     "webpack": "^4.41.6",
     "webpack-cli": "^3.3.11",
-    "webpack-dev-server": "^2.3.0"
+    "webpack-dev-server": "^3.10.3"
   },
   "optionalDependencies": {
     "react": "14 || 15 || 16",
