+++ src/less/themes/goldenlayout-dark-theme.less
@@ -60,7 +60,7 @@
 // Pane Header (container of Tabs for each pane)
 .lm_header {
   height: 20px;
-  
+
   // Single Tab container. A single Tab is set for each pane, a group of Tabs are contained in ".lm_header"
   .lm_tab {
     font-family: Arial, sans-serif;
@@ -81,7 +81,7 @@
     .lm_close_tab {
       width: 11px;
       height: 11px;
-      background-image: data-uri('../img/lm_close_white.png');
+      background-image: data-uri('../../img/lm_close_white.png');
       background-position: center center;
       background-repeat: no-repeat;
       top: 4px;
@@ -165,22 +165,22 @@
 
   // Icon to PopOut Pane, so move it to a different Browser Window
   .lm_popout {
-    background-image: data-uri('../img/lm_popout_white.png');
+    background-image: data-uri('../../img/lm_popout_white.png');
   }
 
   // Icon to Maximize Pane, so it will fill the entire GoldenLayout Container
   .lm_maximise {
-    background-image: data-uri('../img/lm_maximise_white.png');
+    background-image: data-uri('../../img/lm_maximise_white.png');
   }
 
   // Icon to Close Pane and so remove it from GoldenLayout Container
   .lm_close {
-    background-image: data-uri('../img/lm_close_white.png');
+    background-image: data-uri('../../img/lm_close_white.png');
   }
 
-  // Icon to toggle Pane Docking at mouse hover 
+  // Icon to toggle Pane Docking at mouse hover
   .lm_dock {
-    background-image: data-uri('../img/lm_docked_white.png');
+    background-image: data-uri('../../img/lm_docked_white.png');
     transform:rotate(-45deg);
     transition:transform 300ms;
   }
@@ -208,7 +208,7 @@
   // Pane controls are different in Maximized Mode, especially the old Icon "Maximise" that now has a different meaning, so "Minimize" (even if CSS Class did not change)
   .lm_controls {
     .lm_maximise {
-      background-image: data-uri('../img/lm_minimize_white.png');
+      background-image: data-uri('../../img/lm_minimize_white.png');
     }
   }
 }
@@ -230,7 +230,7 @@
 
   // Icon to Restore original position in Golden Layout Container
   .lm_icon {
-    background-image: data-uri('../img/lm_popin_white.png');
+    background-image: data-uri('../../img/lm_popin_white.png');
     background-position: center center;
     background-repeat: no-repeat;
     border-left: 1px solid @color2;
